Delete action is working
App.js
-- Added {match} as a prop to pass to wishlist

Wishlist.js
-- Modified the object that is pushed into the array of data that is
displayed to the user on view load to include the wishlistId. This is
what allows Book.js to target an item for deletion
-- Included 'match' in this.props within the render function
-- Modified book prop to be book={book.bookData} and included
wishlistId={book.wishlistId} as props to the Book component
-- Removed console logs and unneeded comments

Book.js
-- Added "deleted: false" and "read: false" to 'this.state'
-- Modified the axios URL for delete to be `this.props.wishlistId`
-- Created fresh "read" function that will be a button for a user to
click when they have "read" a book that is on their wishlist. This is
still in development and is not working.
-- Included readButton in the same conditional statement as
deleteButton so it only displays when a user is viewing their wishlist.