Added PATCH request for "read"
Book.js
-- Added "data" to markAsRead function's axios call
-- Created markAsUnread function with an axios call, which is basically
the same thing, except setting the status of "read" to "false"
-- Included "read" in the render function
-- Moved the ternary statement determining which button is displayed
from the return up into the conditional statement so that the "markAs"
buttons won't show unless the user is looking at their wishlist.
-- Began an attempt at componentDidMount so that I can set the state
and store it beyond the current view, but it isn't working as yet and is
commented out.